// src/app/Models/JobDetailsById.ts
export interface JobDetailsById {
    id?: any,
    title?: any
    description?: any
    sendMailTo?: any
    imageName?: any
    docName?: any
    pdfName?: any
    active?: boolean
    publishDate?: any
    expiryDate?: any
    createdBy?: any
    uploadFiles?: boolean
    uploadDoc?: boolean,
    uploadPdf?: boolean,
    files?: any
    filesDoc?: any
    filesPdf?: any
    // isUploadFile?: boolean
   
}